GH-46272: [C++] Build Arrow libraries with `-Wmissing-definitions` on gcc (#47042)
### Rationale for this change

The warning option `-Wmissing-declarations` allows finding private functions that erroneously have global linkage (because they are neither static nor in the anonymous namespace).

We only care about this for the public Arrow libraries, not for tests or utilities where it's harmless to have private functions that nevertheless have global linkage (and changing that would require a lot of pointless code churn).

### Are these changes tested?

Yes, on builds using gcc.

### Are there any user-facing changes?

No, because this is only enabled if the warning level is "CHECKIN". Release builds will by default use the "PRODUCTION" warning level.
* GitHub Issue: #46272

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>