GH-41108: [Docs] Remove Sphinx pin (#47326)
### Rationale for this change
Sphinx has been pinned for a while and it is safe to remove the pin now.

### What changes are included in this PR?
Sphinx pin is removed and some minor changes to `exclude_patterns` and an external link are added to reduce couple of build warnings.

### Are these changes tested?
Yes, with CI. The docs build should succeed and the preview build should be run.

### Are there any user-facing changes?
No.
* GitHub Issue: #41108

Authored-by: AlenkaF <frim.alenka@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>