GH-43904: [CI][Python] Stop uploading nightly wheels to gemfury (#47470)
### Rationale for this change

We started uploading nightly wheels to the scientific python channel on anaconda:
https://anaconda.org/scientific-python-nightly-wheels/pyarrow
Which are the documented nightlies:
https://arrow.apache.org/docs/dev/developers/python.html#installing-nightly-packages
We updated the documentation and have been deprecating with a long release cycle.

Uploading wheels to gemfury sometimes fail due to trying to upload the same version of the wheels.

### What changes are included in this PR?

Stop uploading nightly wheels to gemfury.

### Are these changes tested?

No

### Are there any user-facing changes?

Yes, nightly wheels will not be available on gemfury anymore.

* GitHub Issue: #43904

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>