GH-41336: [C++][Compute] Fix case_when kernel dispatch for decimals with different precisions and scales (#47479)
### Rationale for this change

Another case of decimal kernels not able to suppress exact matching when precisions and scales of the arguments differ, causing wrong result type. After #47297, we have a systematic way to do that and guide the matching to go to the "best match" (applying implicit casts).

### What changes are included in this PR?

Simply added a constraint match that checks if the precisions and scales of the decimal arguments are the same. Also added corresponding tests in forms of both expression (exact match first, then best match) and function call (best match only).

### Are these changes tested?

Yes.

### Are there any user-facing changes?

None.
* GitHub Issue: #41336

Authored-by: Rossi Sun <zanmato1984@gmail.com>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>