GH-47591: [C++] Fix passing zlib compression level (#47594)
### Rationale for this change

When passing the compression level to the zlib initiation method `deflateInit2`, we were actually passing it to the wrong parameter (the `memLevel` parameter).

As a consequence, changing the zlib/gzip "compression level" in the Arrow APIs had little effect on actual compressed size.

### What changes are included in this PR?

Pass compression level correctly.

### Are these changes tested?

They are exercised by regular CI tests. In addition, I tested manually that changing the compression level now affects compressed size accordingly.

### Are there any user-facing changes?

Yes, this fixes behavior so as to match the documented semantics.

* GitHub Issue: #47591

Authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Antoine Pitrou <antoine@python.org>